+++ 0015-js-structure/js-structure_quiz-app/index.js
@@ -1,33 +1,7 @@
 import cards from './components/CardList/CardList.js'
 import Header from './components/Header/Header.js'
 import Form from './components/Form/Form.js'
-
-function Bookmark(props) {
-  const bookmark = document.createElement("button");
-  bookmark.classList.add("bookmark");
-
-  if (props.active) {
-    bookmark.classList.add("bookmark--active");
-  }
-
-  bookmark.type = "button";
-  bookmark.setAttribute("aria-label", "bookmark");
-  bookmark.innerHTML = /* html */ `
-		<svg
-			class="bookmark__icon"
-			xmlns="http://www.w3.org/2000/svg"
-			viewbox="0 0 24 24"
-		>
-			<path
-				d="M17 3H7c-1.1 0-2 .9-2 2v16l7-3 7 3V5c0-1.1-.9-2-2-2z"
-			/>
-		</svg>
-	`;
-
-  bookmark.addEventListener("click", props.onClick);
-
-  return bookmark;
-}
+import Bookmark from './components/Bookmark/Bookmark.js'
 
 function Card(props) {
   const card = document.createElement("article");
